Ajout d'une mode de combinaison des grilles. Petites corrections sur les valeurs manquantes (-1, -2 et -3)
git-svn-id: http://rdinnovation.onf.fr/svn/plugin-lvox@30 f1561d03-89af-ab4a-a97c-2bf9dc1f41eb

+++ pluginlvox/step/lvox_stepcomputelvoxgrids.h
@@ -140,7 +140,7 @@ private:
 //********************************************//
 
     double          _res;                   /*!< size of a voxel*/
-    double          _effectiveRayThresh;    /*!< minimum number of effective ray (Nt-Nb) in the voxel to calculate a density*/
+    int             _effectiveRayThresh;    /*!< minimum number of effective ray (Nt-Nb) in the voxel to calculate a density*/
     bool            _computeDistances;
 
 };
